Обновить декодер G.723.1: изменить тип вывода на double, улучшить комментарии и добавить новые заголовочные файлы

+++ g723_1_decoder.h
@@ -10,6 +10,6 @@ void g723_reset_decoder();
 // Декодирование одного фрейма. 
 // Принимает: указатель на входной кадр, массив для записи 240 отсчетов float
 // Возвращает: размер обработанного кадра в байтах (24, 20, 4 или 1 для PLC)
-int g723_decode_frame(const unsigned char* input, float* output_pcm);
+int g723_decode_frame(const unsigned char* input, double* output_pcm);
 
 #endif // G723_1_DECODER_H
